More Markdown tweeking using pureMath/README.md.
Two spaces at end of line and <br> both work to enforce
a line break.  I think using <br> is the right way, since
it explicitly shows my intensions and it is not that much
of a burden when reading the file as text.

I don't like reading lines much greater than 65 characters
in length.