Major progress with fpinscala.chap06.state.candyMachine.
I was gettng confused, I looked at the book answer and
really, really got confused.  My epiphany was when I realized
that the State get and set methods need to be members of the
companion object and not the State case class itself.  Now,
these methods actually make sense to me.

I got the fpinscala chap. 6 candy machine program to compile
and run with a stub simulation (book's simulateMachine) method.