Renamed bash-utils/path to bash-utils/Path, updated comments in bash-utils/whence & linuxHome/bin/whence.
1. Seems that on SuSE Linux (at least a version I was recently on)
the /etc/bash.bashrc script leaves around a shell function
with the path name.
2. My whence script comments refered to the command as Whence.
This was from very long ago when I was on Solaris using ksh.
3. Unlike the ksh built in, my version of whence  only looks
at files along the path, except it does not stop at the first one.
This allows you to see all the versions of programs like cc,
python, or scala on your path.  Also very useful when "helpful"
SAs messed with /etc/bashrc.