One more tweak of fpinscala Par.scala before exercise 7.4.
The Map2Future is "lazy" in the sense that the calculation does not start
until a get or isDone method is called.  When the Map2Future is not in the
"hasStarted" state, I gave the Map2Future a way to avoid the cost of the
function f being calculated even if the underlying af and bf futures can't
be cancelled.  Also, if cancel(false) is called, this future does not even
mess with the underlying af and bf futures when in the "hasStarted" state.