Created fork-par branch to test different sequence implementatons,
Book's implementaton seems to more librally use fork.  In my first
stab I really have to turn up the number of threads to avoid
deadlocks.  This gives more thread overhead.

Time parFilter with with 2 threads: 1287.942 μs.
Time parFilter with with 4 threads: 878.761 μs.
Time parFilter with with 20 threads: 821.268 μs
Time parFilter with with 210 threads: 761.728 μs.
Time parFilter with with 1000 threads: 24405.709 μs.

Time parFilter1 with with 210 threads: 2114.184 μs.
Time parFilter1 with with 1000 threads: 8949.061 μs.

parFilter1 uses a sequence that forks a lot.  Need a
minimum of 200 threads to avoid deadlocking for this
calculation.  Not 100% sure of the story the above
data is trying to tell me, but this deadlocking is a
problem.  The book indicates the root cause of this
deadlocking is the blocking Java Futures used by map2.