Comparisons between different implements of fpinscala.parallelism.Par.sequence with varing degrees of forking.
With current implementations, there doesn't seem much advantage for use with
fairly short calculations,

Time parFilter1 with 350 threads: 37 in 1015.915 μs.
Time parFilter2 with 350 threads: 37 in 2998.86 μs.
Time parFilter3 with 350 threads: 37 in 10095.625 μs.
Time parFilter4 with 350 threads: 37 in 4782.255 μs.
Time parFilter5 with 350 threads: 37 in 4930.342 μs.
Time parFilter  with 350 threads: 37 in 8984.473 μs.
Repeat above without parallelism: 37 in 1027.374 μs.

but with longer running calculation, there seems to be advantages

Time parMap1 with 80 threads: 46 in 4.956093539 s.
Time parMap2 with 80 threads: 46 in 4.930486102 s.
Time parMap3 with 80 threads: 46 in 5.314257319 s.
Time parMap4 with 80 threads: 46 in 5.057085211 s.
Time parMap5 with 80 threads: 46 in 5.37955946 s.
Time parMap  with 80 threads: 46 in 5.216700123 s.
Repeat without parallelism:   46 in 12.547512588 s.