Finally fixed error handling in fpinscala.parallelism.
There is nothing wrong with the Actor implementation borrowed
from the scalaz library.  My code must deal with its own errors;
the onError function that I am passing to the Actor companion
object apply method is for errors internal to an Actor implementation.

I previously removed the default throw(_) implementation for onError
in Actor.scala since throw(_) is an unsafe default in a multithreaded
paradigm since exceptions are too easily swallowed.

Anyway, the only part of the code that uses Actors is map2.