Fixed a bug in package fpinscala.parallelism.javaFutures. This bug made ParJavaFutures1.scala fail and has been around since Aug 12, 2017.
Previously I was using the construct
  new Thread(() => ...).start
where I was taking advantage of Scala's SAM interface.

I had replaced this with
  new Thread {() => ...}.start
which Scala interpreted as an anonymous class extending java's Thread
class.  I am not sure why it even compiled, due to the dangling .start?

For this to work, I would have had to do something like
  val foo = new Thread {
    override def run { ... }
  }
  foo.start

I changed the {  } back to (  ) and my "frozen future" thawed.