First step reorganizing my bookmark webpage.
Goals for this commit are to
1. Make webpage more readable as a text file.
2. Treat each link as a "record."
3. Decomflate table boilerplate from table element layout.
4. Make more "monkey-see-monkey-do" editable.

Most style done in style.css (previously done).  In each table
element "Odd ball" link chosen to be first, not last, since I am
more likely to add a new link to the end of the list.  That is why
the line breaks <br> are before the links.  If I were not exposed
to Haskell layout, I would never have done it that way.